package: Update elm, elm-webpack-loader, but still leaves security vulns.
* elm: 0.18.0 → 0.19.0
* elm-webpack-loader 4.5.0 → 5.0.0

GitHub alerted me to lodash and hoek vulnerabilities based on my
package-lock.json. `npm audit` shows 20 vulnerabilities. This fixes
most of those.

However, it appears to leave extant a high-priority vulnerability in
`tar`[1] due to an elm dependency on binwrap, in turn dependent on a
very old version of tar. Not sure what to do about this, but I don't
think I'm using it.

[1]: https://www.npmjs.com/advisories/803